Исправила названия действий
